requirements.txt

---

### ✅ `requirements.txt`

```txt
selenium
pillow
